Started work on GLSL

+++ src/main/java/me/salamander/ourea/modules/NoiseSampler.java
@@ -1,13 +1,13 @@
 package me.salamander.ourea.modules;
 
 public interface NoiseSampler {
-    void setSalt(long salt);
+    void setSalt(int salt);
     void setFrequency(float frequency);
 
-    float sample(float x, float y, long seed);
-    float sample(float x, float y, float z, long seed);
+    float sample(float x, float y, int seed);
+    float sample(float x, float y, float z, int seed);
 
-    static float safeSample(NoiseSampler sampler, float x, float y, long seed) {
+    static float safeSample(NoiseSampler sampler, float x, float y, int seed) {
         if(sampler == null) {
             return 0;
         }
