+++ src/main/java/me/salamander/ourea/modules/fractal/FBM.java
@@ -46,7 +46,7 @@ public class FBM implements NoiseSampler {
     }
 
     @Override
-    public void setSalt(long salt) {
+    public void setSalt(int salt) {
 
     }
 
@@ -56,7 +56,7 @@ public class FBM implements NoiseSampler {
     }
 
     @Override
-    public float sample(float x, float y, long seed) {
+    public float sample(float x, float y, int seed) {
         float value = 0;
         float frequency = 1;
         float amplitude = 1;
@@ -70,7 +70,7 @@ public class FBM implements NoiseSampler {
     }
 
     @Override
-    public float sample(float x, float y, float z, long seed) {
+    public float sample(float x, float y, float z, int seed) {
         float value = 0;
         float frequency = 1;
         float amplitude = 1;
