+++ src/main/java/me/salamander/ourea/modules/modifier/BinaryModule.java
@@ -13,7 +13,7 @@ public class BinaryModule implements NoiseSampler {
     }
 
     @Override
-    public void setSalt(long salt) {
+    public void setSalt(int salt) {
 
     }
 
@@ -23,12 +23,12 @@ public class BinaryModule implements NoiseSampler {
     }
 
     @Override
-    public float sample(float x, float y, long seed) {
+    public float sample(float x, float y, int seed) {
         return op.apply(first.sample(x, y, seed), second.sample(x, y, seed));
     }
 
     @Override
-    public float sample(float x, float y, float z, long seed) {
+    public float sample(float x, float y, float z, int seed) {
         return op.apply(first.sample(x, y, z, seed), second.sample(x, y, z, seed));
     }
