+++ src/main/java/me/salamander/ourea/modules/source/ValueSampler.java
@@ -4,11 +4,11 @@ import me.salamander.ourea.modules.NoiseSampler;
 import me.salamander.ourea.util.MathHelper;
 
 public class ValueSampler implements NoiseSampler {
-    private long salt;
+    private int salt;
     private float frequency = 1;
 
     @Override
-    public void setSalt(long salt) {
+    public void setSalt(int salt) {
         this.salt = salt;
     }
 
@@ -18,7 +18,7 @@ public class ValueSampler implements NoiseSampler {
     }
 
     @Override
-    public float sample(float x, float y, long seed) {
+    public float sample(float x, float y, int seed) {
         seed += salt;
         x *= frequency;
         y *= frequency;
@@ -39,7 +39,7 @@ public class ValueSampler implements NoiseSampler {
     }
 
     @Override
-    public float sample(float x, float y, float z, long seed) {
+    public float sample(float x, float y, float z, int seed) {
         seed += salt;
         x *= frequency;
         y *= frequency;
